moved journal to it's own folder within page directory
