+++ front-end/src/main.js
@@ -1,5 +1,5 @@
 import { DayComponent } from "./pages/day/DayComponent.js";
-import { JournalComponent } from "./pages/day/journal/JournalComponent.js";
+import { JournalComponent } from "./pages/journal/JournalComponent.js";
 import { EventHub } from "./eventhub/EventHub.js";
 import { Events } from "./eventhub/Events.js";
