+++ front-end/src/pages/journal/JournalComponent.js
@@ -1,18 +1,17 @@
-import { EventHub } from '../../../eventhub/EventHub.js'
-import { Events } from '../../../eventhub/Events.js'
-import { BaseComponent } from '../../main/BaseComponent.js'
+import { EventHub } from '../../eventhub/EventHub.js'
+import { Events } from '../../eventhub/Events.js'
+import { BaseComponent } from '../main/BaseComponent.js'
 
 export class JournalComponent extends BaseComponent {
     constructor() {
-        super('journalPage', './pages/day/journal/stylesJournal.css')
+        super('journalPage', './pages/journal/stylesJournal.css')
         this.dateData = {}
     }
 
 // Methods
     // Returns to Day Page, passes any changes to the date object through an event
     #returnToDayPage() {
-       const hub = EventHub.getInstance()
-       hub.publish(Events.LoadDayPage, this.dateData)
+       EventHub.getInstance().publish(Events.LoadDayPage, this.dateData)
     }
 
     // Stores value in text area to date object
@@ -31,8 +30,8 @@ export class JournalComponent extends BaseComponent {
             <form class="text-submission" id="daySummary">
                 <textarea id="summary" placeholder="Write your summary here (2000 character limit)" maxlength="1000"></textarea>
                 <div class="button-container">
-                    <button id="save" type="button">Save</button>
-                    <button id="cancel" type="button">Cancel</button>
+                    <button id="saveJournal" type="button">Save</button>
+                    <button id="cancelJournal" type="button">Cancel</button>
                 </div>
             </form>
         `   
@@ -41,22 +40,19 @@ export class JournalComponent extends BaseComponent {
     // Adds EventListeners that update attributes in the class
     _addEventListeners() {
         const hub = EventHub.getInstance()
-        hub.subscribe(Events.LoadJournalPage, data => this._render(data))
+        hub.subscribe(Events.LoadJournalPage, data => this.loadPage(data))
 
-        document.getElementById('save').addEventListener('click', () => this.#saveJournal())
-        document.getElementById('cancel').addEventListener('click', () => this.#returnToDayPage())
+        document.getElementById('saveJournal').addEventListener('click', () => this.#saveJournal())
+        document.getElementById('cancelJournal').addEventListener('click', () => this.#returnToDayPage())
     }
 
     // Changes the current view to the Journal Page
-    _render(date_entry) {
-        document.querySelectorAll('.view').forEach(body => body.style.display = 'none')
-        this.dateData = date_entry
+    _render(data) {
+        this.dateData = data
 
         document.getElementById('summary').value = 'journal_entry' in this.dateData
         ? this.dateData.journal_entry
         : ''
-        
-        this._changeDisplay('flex')
     }
 
 }
